Refactor to modularize different react components.  Start work on Treasure Hunts List component.
