+++ app/components/common/MyStatusBar.js
@@ -1,0 +1,22 @@
+'use strict'
+
+import React, { Component } from 'react';
+import {
+  StyleSheet,
+  View,
+  StatusBar,
+} from 'react-native';
+
+var MyStatusBar = ({backgroundColor}) => (
+  <View style={[styles.statusBar, { backgroundColor }]}>
+    <StatusBar backgroundColor={backgroundColor} barStyle="light-content" />
+  </View>
+);
+
+const styles = StyleSheet.create({
+  statusBar: {
+    height: 20,
+  },
+});
+
+module.exports = MyStatusBar;
