+++ app/components/common/sideMenu/TreasureHuntList/ViewContainer.js
@@ -1,0 +1,24 @@
+'use strict'
+
+import { Component, View } from 'react-native'
+
+class ViewContainer extends Component {
+  render() {
+    return (
+      <View style={styles.viewContainer}>
+        {this.props.children}
+      </View>
+    )
+  }
+}
+
+const styles = React.StyleSheet.create({
+
+  viewContainer: {
+    flex: 1,
+    flexDirection: 'column',
+    justifyContent: 'flex-start',
+    alignItems: 'stretch'
+  }
+
+})
