+++ app/components/common/sideMenu/TreasureHuntList/treasureHuntList.js
@@ -1,0 +1,9 @@
+'use strict'
+
+
+class TreasureHuntList extends Component {
+  render() {
+    <ViewContainer>
+    </ViewContainer>
+  }
+}
